fixed unintentional rebounds
Happened when holding in the jump key from a high jump and hitting the ground.  It didn't matter if the mode had rebound set or not.  Flight was continuing and something was adding an impulse when it shouldn't have

This detects that scenario and explicitly tells keydash_tracker to stop until they release jump, hold jump